torch_methods/is_cuda_avaliable.py
import torch


def check_cuda():
    print("Torch version:", torch.__version__)
    print("CUDA available:", torch.cuda.is_available())
    print("CUDA version:", torch.version.cuda)

    if torch.cuda.is_available():
        print("GPU:", torch.cuda.get_device_name(0))
    else:
        print("Using CPU")

if __name__ == "__main__":
    check_cuda()